requirements.txt
PyGithub==1.25.2
argparse==1.2.1
pymongo==3.0.1
python-dateutil==2.4.2
six==1.9.0
wsgiref==0.1.2
